TIP-874: move Channel constraints
