+++ src/Akeneo/Channel/Bundle/DependencyInjection/AkeneoChannelExtension.php
@@ -31,6 +31,7 @@ class AkeneoChannelExtension extends Extension
         $loader->load('job_constraints.yml');
         $loader->load('services.yml');
         $loader->load('steps.yml');
+        $loader->load('validators.yml');
         $loader->load('writers.yml');
     }
 }
