+++ src/Akeneo/Channel/Component/Validator/Constraint/IsRootCategoryValidator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\Catalog\Validator\Constraints;
+namespace Akeneo\Channel\Component\Validator\Constraint;
 
 use Akeneo\Pim\Enrichment\Component\Category\Model\CategoryInterface;
 use Symfony\Component\Validator\Constraint;
