+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/validators.yml
@@ -7,9 +7,6 @@ parameters:
     pim_catalog.validator.constraint.scopable_value.class:                 Pim\Component\Catalog\Validator\Constraints\ScopableValueValidator
     pim_catalog.validator.constraint.localizable_value.class:              Pim\Component\Catalog\Validator\Constraints\LocalizableValueValidator
     pim_catalog.validator.constraints.channel_validator.class:             Pim\Component\Catalog\Validator\Constraints\ChannelValidator
-    pim_catalog.validator.constraints.is_root_category.class:              Pim\Component\Catalog\Validator\Constraints\IsRootCategoryValidator
-    pim_catalog.validator.constraints.conversion_units.class:              Pim\Component\Catalog\Validator\Constraints\ConversionUnitsValidator
-    pim_catalog.validator.constraints.is_currency_activated.class:         Pim\Component\Catalog\Validator\Constraints\IsCurrencyActivatedValidator
     pim_catalog.validator.constraint_guesser.chained.class:                Pim\Component\Catalog\Validator\ChainedAttributeConstraintGuesser
     pim_catalog.validator.constraint_guesser.email.class:                  Pim\Component\Catalog\Validator\ConstraintGuesser\EmailGuesser
     pim_catalog.validator.constraint_guesser.file.class:                   Pim\Component\Catalog\Validator\ConstraintGuesser\FileGuesser
@@ -135,24 +132,6 @@ services:
             - '@pim_catalog.repository.cached_locale'
         tags:
             - { name: validator.constraint_validator, alias: pim_localizable_value_validator }
-
-    pim_catalog.validator.constraints.is_root_category:
-        class: '%pim_catalog.validator.constraints.is_root_category.class%'
-        tags:
-            - { name: validator.constraint_validator, alias: pim_is_root_category_validator }
-
-    pim_catalog.validator.constraints.conversion_units:
-        class: '%pim_catalog.validator.constraints.conversion_units.class%'
-        arguments:
-            - '@pim_catalog.repository.attribute'
-            - '@akeneo_measure.manager'
-        tags:
-            - { name: validator.constraint_validator, alias: pim_conversion_units_validator }
-
-    pim_catalog.validator.constraints.is_currency_activated:
-        class: '%pim_catalog.validator.constraints.is_currency_activated.class%'
-        tags:
-            - { name: validator.constraint_validator, alias: pim_is_currency_activated_validator }
 
     # Attribute constraint guesser
     pim_catalog.validator.constraint_guesser.chained_attribute:
